Refactor packages and optimized loading movies from the network in SplashScreenActivity
