+++ app/src/main/AndroidManifest.xml
@@ -11,21 +11,20 @@
         android:roundIcon="@drawable/cinema"
         android:supportsRtl="true"
         android:theme="@style/Theme.DoboshAcademyApp">
-        <activity android:name=".view.MainActivity">
+        <activity android:name=".views.activities.MainActivity"/>
 
-        </activity>
-        <activity android:name=".view.SplashScreenActivity"
-            android:theme="@style/SplashScreenTheme"
-            >
+        <activity android:name=".views.activities.SplashScreenActivity"
+            android:theme="@style/SplashScreenTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
+
     </application>
 
 </manifest>
