+++ app/src/main/java/com/johnsondev/doboshacademyapp/network/api/MovieApi.kt
@@ -1,20 +1,20 @@
 package com.johnsondev.doboshacademyapp.network.api
 
-import com.johnsondev.doboshacademyapp.model.response.MovieTopRatedResponse
-import com.johnsondev.doboshacademyapp.model.dto.MovieDto
-import com.johnsondev.doboshacademyapp.model.response.CastResponse
-import com.johnsondev.doboshacademyapp.model.response.MoviePopularResponse
-import com.johnsondev.doboshacademyapp.model.response.MovieUpcomingResponse
+import com.johnsondev.doboshacademyapp.data.dto.MovieDto
+import com.johnsondev.doboshacademyapp.data.response.*
 import retrofit2.http.GET
 import retrofit2.http.Path
 
 interface MovieApi {
 
     @GET("movie/top_rated")
-    suspend fun getTopRated(): MovieTopRatedResponse
+    suspend fun getTopRated(): MovieResponse
 
     @GET("movie/popular")
-    suspend fun getPopular(): MoviePopularResponse
+    suspend fun getPopular(): MovieResponse
+
+    @GET("movie/upcoming")
+    suspend fun getUpcoming(): MovieResponse
 
     @GET("movie/{movie_id}")
     suspend fun getMovieById(@Path(value = "movie_id") id: Int): MovieDto
@@ -22,7 +22,6 @@ interface MovieApi {
     @GET("movie/{movie_id}/credits")
     suspend fun getActors(@Path(value = "movie_id") id: Int): CastResponse
 
-    @GET("movie/upcoming")
-    suspend fun getUpcoming(): MovieUpcomingResponse
+
 
 }
