+++ app/src/main/java/com/johnsondev/doboshacademyapp/utilities/DtoMapper.kt
@@ -1,13 +1,13 @@
-package com.johnsondev.doboshacademyapp.tools
+package com.johnsondev.doboshacademyapp.utilities
 
-import com.johnsondev.doboshacademyapp.model.data.Actor
-import com.johnsondev.doboshacademyapp.model.data.Genre
-import com.johnsondev.doboshacademyapp.model.data.Movie
-import com.johnsondev.doboshacademyapp.model.dto.ActorDto
-import com.johnsondev.doboshacademyapp.model.dto.GenreDto
-import com.johnsondev.doboshacademyapp.model.dto.MovieDto
-import com.johnsondev.doboshacademyapp.tools.Constants.ACTOR_IMG_PATH
-import com.johnsondev.doboshacademyapp.tools.Constants.POSTER_PATH
+import com.johnsondev.doboshacademyapp.data.models.Actor
+import com.johnsondev.doboshacademyapp.data.models.Genre
+import com.johnsondev.doboshacademyapp.data.models.Movie
+import com.johnsondev.doboshacademyapp.data.dto.ActorDto
+import com.johnsondev.doboshacademyapp.data.dto.GenreDto
+import com.johnsondev.doboshacademyapp.data.dto.MovieDto
+import com.johnsondev.doboshacademyapp.utilities.Constants.ACTOR_IMG_PATH
+import com.johnsondev.doboshacademyapp.utilities.Constants.POSTER_PATH
 
 object DtoMapper {
