+++ app/src/main/java/com/johnsondev/doboshacademyapp/views/activities/MainActivity.kt
@@ -1,9 +1,9 @@
-package com.johnsondev.doboshacademyapp.view
+package com.johnsondev.doboshacademyapp.views.activities
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.johnsondev.doboshacademyapp.R
-import com.johnsondev.doboshacademyapp.view.movielist.FragmentMoviesList
+import com.johnsondev.doboshacademyapp.views.movielist.FragmentMoviesList
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
